config/urls.py

from django.contrib import admin
from django.urls import include, path
from rest_framework.routers import DefaultRouter
from alunos.views import EstadoViewSet, CidadeViewSet, PessoaViewSet

router = DefaultRouter()
router.register(r'estados', EstadoViewSet)
router.register(r'cidades', CidadeViewSet)
router.register(r'pessoas', PessoaViewSet)




urlpatterns = [
    path('admin/', admin.site.urls),
    path('', include(router.urls))
]
